Move streaming picker primitives into ui.interactive_picker
Turn ui.py into a ui/ package and move the streaming multi-select picker primitives (scrolling_checkbox, StreamingInquirerControl, merge_new_choices, picker_style, PICKER_VISIBLE_ROWS) out of mcp.py into ui/interactive_picker.py, re-exported from ucode.ui. Add a loading_noun parameter so non-MCP pickers can label the loading footer. No behavior change; picker unit tests move to tests/test_interactive_picker.py.
